using intptr_t instead of ulong/void* based custom defn.

+++ test/id_mgr_test.c
@@ -15,7 +15,7 @@ void test_create_query(void)
     printf("%s: ", __func__);
     libid_init(UT_GC_TIMEOUT);
 
-    uint32 dt1_id = create_id(dt1);
+    uint32 dt1_id = create_id((intptr_t) dt1);
     assert(dt1_id == 1);
 
     char* dt1_t = (char *)query_id(dt1_id);
@@ -32,19 +32,19 @@ void test_create_delete_query(void)
     
     libid_init(UT_GC_TIMEOUT);
 
-    uint32 dt1_id = create_id(dt1);
+    uint32 dt1_id = create_id((intptr_t) dt1);
     assert(dt1_id == 1);
 
-    uint32 dt1_id_t = create_id(dt1);
+    uint32 dt1_id_t = create_id((intptr_t) dt1);
     assert(dt1_id == dt1_id_t);
 
     char* dt1_t = (char *)query_id(dt1_id);
     assert(strcmp(dt1_t, dt1) == 0);
 
-    uint32 dt2_id = create_id(dt2);
+    uint32 dt2_id = create_id((intptr_t) dt2);
     assert(dt2_id == 2);
 
-    uint32 dt3_id = create_id(dt3);
+    uint32 dt3_id = create_id((intptr_t) dt3);
     assert(dt3_id == 3);
 
     delete_id(dt1_id);
@@ -72,20 +72,20 @@ void test_create_delete_create(void)
     printf("%s: ", __func__);
     libid_init(UT_GC_TIMEOUT);
 
-    uint32 dt1_id = create_id(dt1);
-    uint32 dt2_id = create_id(dt2);
-    uint32 dt3_id = create_id(dt3);
-    uint32 dt4_id = create_id(dt4);
+    uint32 dt1_id = create_id((intptr_t) dt1);
+    uint32 dt2_id = create_id((intptr_t) dt2);
+    uint32 dt3_id = create_id((intptr_t) dt3);
+    uint32 dt4_id = create_id((intptr_t) dt4);
 
     delete_id(dt1_id);
     delete_id(dt2_id);
     delete_id(dt3_id);
     delete_id(dt4_id);
 
-    uint32 dt1_id_n = create_id(dt1);
-    uint32 dt2_id_n = create_id(dt2);
-    uint32 dt3_id_n = create_id(dt3);
-    uint32 dt4_id_n = create_id(dt4);
+    uint32 dt1_id_n = create_id((intptr_t) dt1);
+    uint32 dt2_id_n = create_id((intptr_t) dt2);
+    uint32 dt3_id_n = create_id((intptr_t) dt3);
+    uint32 dt4_id_n = create_id((intptr_t) dt4);
     assert(dt1_id != dt1_id_n);
     assert(dt2_id != dt2_id_n);
     assert(dt3_id != dt3_id_n);
@@ -103,10 +103,10 @@ void test_create_delete_gc_create(void)
     printf("%s: ", __func__);
     libid_init(UT_GC_TIMEOUT);
 
-    uint32 dt1_id = create_id(dt1);
-    uint32 dt2_id = create_id(dt2);
-    uint32 dt3_id = create_id(dt3);
-    uint32 dt4_id = create_id(dt4);
+    uint32 dt1_id = create_id((intptr_t) dt1);
+    uint32 dt2_id = create_id((intptr_t) dt2);
+    uint32 dt3_id = create_id((intptr_t) dt3);
+    uint32 dt4_id = create_id((intptr_t) dt4);
 
     delete_id(dt1_id);
     delete_id(dt2_id);
@@ -115,10 +115,10 @@ void test_create_delete_gc_create(void)
 
     sleep(UT_GC_TIMEOUT + 1);
 
-    uint32 dt1_id_n = create_id(dt1);
-    uint32 dt2_id_n = create_id(dt2);
-    uint32 dt3_id_n = create_id(dt3);
-    uint32 dt4_id_n = create_id(dt4);
+    uint32 dt1_id_n = create_id((intptr_t) dt1);
+    uint32 dt2_id_n = create_id((intptr_t) dt2);
+    uint32 dt3_id_n = create_id((intptr_t) dt3);
+    uint32 dt4_id_n = create_id((intptr_t) dt4);
     assert(dt1_id == dt1_id_n);
     assert(dt2_id == dt2_id_n);
     assert(dt3_id == dt3_id_n);
@@ -137,22 +137,22 @@ void test_create_ref_count_delete_gc_create(void)
     libid_init(UT_GC_TIMEOUT);
 
     // ref-count=1
-    uint32 dt1_id = create_id(dt1);
-    uint32 dt2_id = create_id(dt2);
-    uint32 dt3_id = create_id(dt3);
-    uint32 dt4_id = create_id(dt4);
+    uint32 dt1_id = create_id((intptr_t) dt1);
+    uint32 dt2_id = create_id((intptr_t) dt2);
+    uint32 dt3_id = create_id((intptr_t) dt3);
+    uint32 dt4_id = create_id((intptr_t) dt4);
 
     // ref-count=2
-    assert(dt1_id == create_id(dt1));
-    assert(dt2_id == create_id(dt2));
+    assert(dt1_id == create_id((intptr_t) dt1));
+    assert(dt2_id == create_id((intptr_t) dt2));
 
     // ref-count=3
-    assert(dt1_id == create_id(dt1));
-    assert(dt2_id == create_id(dt2));
+    assert(dt1_id == create_id((intptr_t) dt1));
+    assert(dt2_id == create_id((intptr_t) dt2));
 
     // ref-count=4
-    assert(dt1_id == create_id(dt1));
-    assert(dt2_id == create_id(dt2));
+    assert(dt1_id == create_id((intptr_t) dt1));
+    assert(dt2_id == create_id((intptr_t) dt2));
 
     delete_id(dt1_id); // ref-count=3
     delete_id(dt2_id); // ref-count=3
@@ -168,8 +168,8 @@ void test_create_ref_count_delete_gc_create(void)
     delete_id(dt2_id); // ref-count=0
     sleep(UT_GC_TIMEOUT + 1);
 
-    uint32 dt1_id_n = create_id(dt1); // ref-count=2
-    uint32 dt2_id_n = create_id(dt2); // ref-count=1 (re-use)
+    uint32 dt1_id_n = create_id((intptr_t) dt1); // ref-count=2
+    uint32 dt2_id_n = create_id((intptr_t) dt2); // ref-count=1 (re-use)
     assert(dt1_id == dt1_id_n);
     assert(dt2_id == dt2_id_n);
     printf(" PASSED\n");
